Fix some compiler warnings, add more clean logs in merge notification handler Add refresh helper

+++ MagicalRecord/Categories/DataImport/NSAttributeDescription+MagicalDataImport.m
@@ -42,7 +42,7 @@
                 do {
                     NSMutableString *dateFormatKey = [kMagicalRecordImportCustomDateFormatKey mutableCopy];
                     if (index) {
-                        [dateFormatKey appendFormat:@".%d", index];
+                        [dateFormatKey appendFormat:@".%lu", (unsigned long)index];
                     }
                     index++;
                     dateFormat = [[self userInfo] valueForKey:dateFormatKey];
